==> RNAseq_Analysis/Differential_Editing/SPRINT/4_merge_filter_sites.ipynb <==
import os
import pandas as pd 

# Provide the directory path
directory_path = '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/5dpf'

# Create an empty DataFrame to store data
merged_df = pd.DataFrame()

# Find all files in the directory with the "_editing.txt" suffix
file_names = [file for file in os.listdir(directory_path) if file.endswith("_editing.txt")]

# Process each file
for file in file_names:
    # Extract sample name from the file
    sample_name = file.replace("_editing.txt", "")

    # Read data from the file
    file_path = os.path.join(directory_path, file)
    dtype_dict = {'chr': str, 'start': int, 'stop': int, 'editing type': str, 'read type': str, 'strand': str, 'coverage': str}
    df = pd.read_csv(file_path, sep='\t', dtype=dtype_dict)


    # Rename the 'coverage' column to include the sample name
    df.rename(columns={'coverage': f'{sample_name} coverage'}, inplace=True)

    # Merge the data into the main DataFrame
    if merged_df.empty:
        merged_df = df
    else:
        merged_df = pd.merge(merged_df, df, on=['chr', 'start', 'stop', 'editing type', 'read type', 'strand'], how='outer', suffixes=('', f'_{sample_name}'))

# Display the intermediate DataFrame
merged_df.head(100)

# Fill NaN values with '.'
merged_df.fillna(0, inplace=True)

# Display the updated DataFrame
merged_df.head()

merged_df

coverage_columns = ["Ctrl-04 coverage", "Ctrl-05 coverage", "Ctrl-06 coverage",  "NO-04 coverage", "NO-05 coverage", "NO-06 coverage"
]

# Create new columns for each sample to store the summed coverage (A + B)
for col in coverage_columns:
    merged_df[f"{col}_sum"] = merged_df[col].apply(lambda x: sum(map(int, str(x).split(":"))) if ':' in str(x) else int(x) if str(x) != 'nan' else 0)

merged_df

# Check if all samples have a minimum coverage of 10 for each row
acceptable_rows = merged_df.apply(lambda row: all(row[f"{col}_sum"] >= 10 for col in coverage_columns), axis=1)

# Filter the rows
filtered_df = merged_df.loc[acceptable_rows].copy()  # Use .loc to avoid chained indexing

filtered_df

# Calculate the percentage of edited sites for each sample
for col in coverage_columns:
    filtered_df[f"{col}_editing_percentage"] = filtered_df.apply(lambda row: (int(row[col].split(":")[1]) / (int(row[f"{col}_sum"]) )) * 100 if ':' in str(row[f"{col}"]) else 0, axis=1)

filtered_df

filtered_df = filtered_df.drop(columns=[f"{col}_sum" for col in coverage_columns], errors='ignore')


filtered_df

# Save the merged data to a new file
output_filename = os.path.join(directory_path, "SPRINT_ZFR_editing_5dpf.tsv")
filtered_df.to_csv(output_filename, sep='\t', index=False)

print(f"Merged data written to {output_filename}")
