==> RNAseq_Analysis/Differential_Editing/SPRINT/5_prep_REDIT_input.ipynb <==
### Prepare RNA Editing data from SPRINT for REDIT Statistical Inference
##### The purpose of this notebook is to manipulate the data into the proper format to run REDIT R package that tests the statistical inference of RNA editing sites. For each site, the format required by REDIT is as follows:

|                       | Mutant | Mutant | Mutant | Control | Control | Control |
|-----------------------|--------|--------|--------|---------|---------|---------|
| Number of edited reads     |        |        |        |         |         |         |
| Number of non-edited reads |        |        |        |         |         |         |

##### We will also create an ID column with the chr_site (eg. 1_209063) to identify the site at which the editing is occuring and append that on after running the REDIT package but for now we will just create the proper format to easily extract the data in R.

import pandas as pd

df=pd.read_csv("/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/5dpf/SPRINT_ZFR_editing_5dpf.tsv",sep='\t')

#First let's get rid of any data we won't be using
df

# Filter rows based on the "chr" column
filtered_df = df[df['chr'].astype(str).str.match(r'^[1-9]|1[0-9]|2[0-5]$')]

# Create a new DataFrame with the selected columns
new_df = filtered_df.iloc[:, [0, 2] + list(range(6, 12))]

new_df

# Create the "ID" column
id_column = new_df['chr'].astype(str) + '_' + new_df['stop'].astype(str)

# Create a DataFrame for the "ID" column
id_matrix = pd.DataFrame({'ID': id_column})

# Display the "ID" matrix
print(id_matrix)

# Define a function to split, rearrange, and swap the values
def split_rearrange_and_swap(row):
    # Split the values in the coverage columns by ':'
    split_values = [value.split(':') for value in row[2:]]
    
    # Swap the values
    swapped_values = [(value[1], value[0]) for value in split_values]
    
    # Transpose the result to have separate rows for 'Edited' and 'Non-Edited'
    edited_nonedited_values = list(zip(*swapped_values))
    
    # Create a DataFrame for the result
    result_df = pd.DataFrame(edited_nonedited_values, index=['Edited', 'Non-Edited'], columns=new_df.columns[2:])
    
    return result_df

# Apply the function to each row
result_matrices = new_df.apply(split_rearrange_and_swap, axis=1)

# Concatenate the list of DataFrames into a single DataFrame
result_df = pd.concat(result_matrices.tolist(), keys=id_matrix['ID'], axis=0)

# Convert the index values to strings and include "Edited" and "Non-Edited"
result_df.index = result_df.index.map(lambda x: f"{str(x[0])} {x[1]}")

# Display the final result
print(result_df)

result_df.to_csv('/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/5dpf/5dpf_REDIT_input.tsv', sep='\t')
