==> RNAseq_Analysis/Differential_Editing/SPRINT/6_REDIT_Regression.ipynb <==
library(doParallel)
library(foreach)
source("/home/hcs2152/github/REDITs/REDIT_regression.R")

# Input data 
file_2dpf <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/2dpf/2dpf_REDIT_input.tsv'
file_5dpf <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/5dpf/5dpf_REDIT_input.tsv'

# Read the TSV file into a data frame
data_2dpf <- read.table(file_2dpf, header = TRUE, sep = "\t")
data_5dpf <- read.table(file_5dpf, header = TRUE, sep = "\t")

merged_df <- merge(data_2dpf, data_5dpf, by = "X", all = FALSE)
merged_df

# Now construct the coviariate matrix as described by REDIT tutorial 
the_covariates = data.frame(group=c("control", "control", "control", "mutant", "mutant", "mutant", "control", "control", "control", "mutant", "mutant", "mutant"), dpf=c(2,2,2,2,2,2,5,5,5,5,5,5))
the_covariates

# Extract the first column (index column)
ids <- merged_df[, 1]

# Exclude the first column for further processing
merged_df <- merged_df[, -1]

# Create a new data frame with the extracted part before the space (the site)
new_ids <- data.frame(ID = sub(" .*", "", ids))
# Remove duplicates since we had 2 rows for each site (edited/non-edited)
new_ids <- unique(new_ids)

#Split up the table into separate matrices for processing
rows_per_matrix <- 2

matrix_list <- split(merged_df, (seq(nrow(merged_df))-1) %/% rows_per_matrix)

# Assuming matrix_list is a list of data frames
matrix_list <- lapply(matrix_list, as.matrix)

#initiate a cluster
noCores = detectCores() -1
cl = makeCluster(noCores,outfile="")
registerDoParallel(cl,cores=noCores)

length(matrix_list)

output_matrix = foreach(i=1:length(matrix_list),.combine='rbind') %dopar%{
    current_matrix <- matrix_list[[i]]
    regression_info = REDIT_regression(data=current_matrix, covariates=the_covariates)
    return( as.matrix( data.frame(p_value= regression_info$dpf.p.value )) )
}

output_matrix

p_vals_per_site <- cbind(new_ids,output_matrix)

p_vals_per_site$p_adj_BH <- p.adjust(p_vals_per_site$p_value, method = "BH")

p_vals_per_site

p_file <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/SPRINT_combined_p_values.tsv'

write.table(p_vals_per_site, file = p_file, sep = "\t", quote = FALSE, row.names = FALSE)
