==> RNAseq_Analysis/Differential_Editing/SPRINT/7_merge_outputs.ipynb <==
# Input data 
file_2dpf <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/2dpf/SPRINT_ZFR_editing_2dpf.tsv'
file_5dpf <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/5dpf/SPRINT_ZFR_editing_5dpf.tsv'
file_p_vals <- ('/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/SPRINT_combined_p_values.tsv')
# Read the TSV file into a data frame
data_2dpf <- read.table(file_2dpf, header = TRUE, sep = "\t")
data_5dpf <- read.table(file_5dpf, header = TRUE, sep = "\t")
data_p_vals <- read.table(file_p_vals, header = TRUE, sep = "\t")

#Remove any contigs besides Chr 1-25
pattern <- "^(1[0-9]|2[0-5]|[1-9])$"
data_2dpf <- subset(data_2dpf, grepl(pattern, chr))
data_5dpf <- subset(data_5dpf, grepl(pattern, chr))

#Remove the start column (site of editing is stop column)
data_2dpf <- subset(data_2dpf, select = -start)
data_5dpf <- subset(data_5dpf, select = -start)


# Create 'ID' column by pasting 'chr' and 'stop' with '_'
data_2dpf$ID <- paste(data_2dpf$chr, data_2dpf$stop, sep = "_")
data_5dpf$ID <- paste(data_5dpf$chr, data_5dpf$stop, sep = "_")
# Rearrange columns with 'ID' as the first column and remove 'chr' and 'stop' columns
data_2dpf <- data_2dpf[, c("ID", setdiff(names(data_2dpf), c("ID", "chr", "stop")))]
data_5dpf <- data_5dpf[, c("ID", setdiff(names(data_5dpf), c("ID", "chr", "stop")))]

merged_df <- merge(data_2dpf, data_5dpf, by = "ID")

#Add in p and p adj
final_df <- merge(merged_df, data_p_vals, by = "ID")

final_df

file_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/SPRINT_zfr_combined_final.tsv'

# Export the merged data frame to a TSV file
write.table(final_df, file = file_path, sep = "\t", row.names = FALSE)
