==> RNAseq_Analysis/Differential_Editing/SPRINT/8_remove_dpSNPs.ipynb <==
Remove any dbSNPs and rerun the multiple testing correction

dbsnp_file <- read.table('/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/References/Danio_rerio_dbSNP.txt',header=TRUE,sep='\t')
editing_sites <- read.table('/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/SPRINT_zfr_combined_final.tsv',header=TRUE,sep='\t')

#put dbSNP table into ID format like editing sites
dbsnp_file$ID <- paste(dbsnp_file$genomic.region, dbsnp_file$coordinate, sep = '_')

# Move 'ID' column to the beginning
dbsnp_file <- dbsnp_file[, c('ID', names(dbsnp_file)[-which(names(dbsnp_file) %in% 'ID')])]

dbsnp_file

library(dplyr)

#Note the number of editing sites before
dim(editing_sites)

# Anti-join to remove matching rows from editing_sites
editing_sites <- anti_join(editing_sites, dbsnp_file, by = 'ID')

#Check to see if/how many were removed
dim(editing_sites)

#Redo the multiple testing correction since there are a different number of data points now
#Remove old p-adj bh values
editing_sites <- editing_sites[, !names(editing_sites) %in% 'p_adj_BH']

editing_sites$p_adj_BH <- p.adjust(editing_sites$p_value, method = "BH")

editing_sites

significant_ids <- editing_sites$ID[editing_sites$p_adj_BH < 0.05]

significant_ids

p_file <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/SPRINT/Output/A2I_Editing/SPRINT_ZFR_combined_dbSNP_removed_final.tsv'
write.table(editing_sites, file = p_file, sep = "\t", quote = FALSE, row.names = FALSE)
